Eric Sunshine applied patch from Mike Gist <xordan@gmail.com> which works around crasher in csImage::Rescale() on AMD64. Very possibly, the problem was result of unintended sign-extension when assigning signed int csQint() result to unsigned long.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@19204 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csgfx/csimage.cpp
@@ -190,9 +190,9 @@ void csImageFile::Rescale (int newwidth, int newheight)
   // such as blending multiple pixels together or bilinear filtering,
   // just a rough scale. It could be improved by someone in the future.
 
-  unsigned long x, y;
-  unsigned long dx = csQint16 (float (Width) / float (newwidth));
-  unsigned long dy = csQint16 (float (Height) / float (newheight));
+  unsigned int x, y;
+  unsigned int dx = csQint16 (float (Width) / float (newwidth));
+  unsigned int dy = csQint16 (float (Height) / float (newheight));
 
 #define RESIZE(pt, field)				\
   {							\
